split fields into new component
