+++ src/components/DatexBlockSection.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { onBeforeUnmount, onMounted, ref } from 'vue';
+import DatexBlockField from '@/components/DatexBlockField.vue';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const props = defineProps({
@@ -8,8 +9,6 @@ const props = defineProps({
     required: true,
   },
 });
-
-const bytesCutoff: number = 30;
 
 const box = ref<HTMLDivElement | null>(null);
 const width = ref(0);
@@ -52,60 +51,21 @@ function getChUnitInElement(el: Element) {
   el.removeChild(test);
   return chWidth;
 }
-
-const uint8ToHexString = (b: number): string => b.toString(16).padStart(2, '0');
-
-// this will later hopefully be replaced by shadcn Tooltip hovering
-// document.querySelectorAll('.field-wrapper').forEach((e, i) => {
-//   e.addEventListener('mousemove', () => console.log(i))
-// })
-
-const categories = [
-  ['Magic Number', 'Version', 'Context ID', 'Section Index', 'Block Number'],
-  ['Block Size', 'Flags', 'Flags and Timestamp', 'Lifetime'],
-  ['Checksum', 'Signature', 'Encrypted Signature', 'IV'],
-  ['Distance', 'TTL'],
-  ['Sender', 'Number of Receivers', 'Receivers', 'Represented By', 'On Behalf Of'],
-  ['Receivers Pointer ID', 'Receivers with Keys'],
-];
-
-function getCategoryColor(str: string) {
-  const index = categories.findIndex((subArray) => subArray.includes(str));
-  return `var(--color-chart-${index !== -1 ? (index % 5) + 1 : null}`;
-}
-
 </script>
 
 <template>
   <div class="block-protocol-section">
     <div ref="box" :style="`grid-template-columns: repeat(auto-fit, 3.5ch);`" class="bytes-grid">
-      <div
-        class="field-wrapper"
+      <DatexBlockField
         v-for="(field, indexOuter) in section.fields"
-        @click="console.log(`Name: ${field.name}, Value: ${field.parsedValue}`)"
         :key="indexOuter"
-      >
-        <div
-          v-for="(byte, indexInner) in field.bytes.slice(0, bytesCutoff)"
-          :key="indexInner"
-          :style="{ backgroundColor: getCategoryColor(field.name) }"
-        >
-          {{ uint8ToHexString(byte) }}
-        </div>
-        <div
-          v-if="field.bytes.length > bytesCutoff"
-          :style="{ backgroundColor: getCategoryColor(field.name) }"
-        >
-          ..
-        </div>
-      </div>
+        :field="field"
+        :indexOuter="indexOuter"
+      ></DatexBlockField>
     </div>
-    <!-- <div style="background-color: green">Width: {{ width }}px</div>
-    <div>Bytes that fit: {{ bytes }}</div> -->
   </div>
 </template>
 
-<!-- lang="postcss" um tailwind zu benutzen-->
 <style scoped>
 .block-protocol-section {
   background-color: var(--color-secondary);
@@ -122,28 +82,5 @@ function getCategoryColor(str: string) {
   grid-row-gap: 0.75ch;
   font-family: monospace;
   width: 100%;
-}
-
-.field-wrapper {
-  display: contents;
-}
-
-.field-wrapper div {
-  padding-left: 0.75ch;
-}
-
-/* nochmal genau die margins und paddings prüfen */
-.field-wrapper div:first-child {
-  border-bottom-left-radius: var(--radius-sm);
-  border-top-left-radius: var(--radius-sm);
-  margin-left: 0.25ch;
-  padding-left: 0.5ch;
-}
-
-.field-wrapper div:last-child {
-  border-bottom-right-radius: var(--radius-sm);
-  border-top-right-radius: var(--radius-sm);
-  margin-right: 0.25ch;
-  padding-right: 0.25ch;
 }
 </style>
